backend/requirements.txt
arrow==1.3.0
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
h11==0.14.0
idna==3.6
numpy==1.26.4
outcome==1.3.0.post0
pandas==2.2.1
psycopg2-binary==2.9.9
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyVirtualDisplay==3.0
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
types-python-dateutil==2.8.19.20240311
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wsproto==1.2.0
